Created brainstorming file
I added a text doc for us to brainstorm ideas about enemies, power-ups, and game mechanics. This isn't super urgent yet, but if you have any ideas, add them here. It's better to have too many ideas and whittle them down (like if we decide we don't like them or they're too hard to figure out) than to be stuck with nothing and try to come up with something. If you think one of my ideas sucks, feel free to change it.