Worked on weapons and whitecaps
Bullets now get killed when they hit islands (sorry Dan), and there is code to limit the number of shots the player can take based on the number of kills they have. Still need to figure out how to reset the shots after a delay. Also tried (and failed) to get whitecaps to wrap around the world and t odie when they hit the ship or islands