Almost got tack sprites working
I got rid of a ton of now-unused code, too. For some reason, the ship likes to be in port tack more than it should...that's the next thing to fix I also did some brainstorming, coming up wit hthe ironclad and ghost ship bosses.
We may want to have an "upwind" sprite on the spritesheet to get rid of some of the ambiguity of the upwind situation