Shooting kills enemies
Still a long way to go on getting waves to work properly, and the game isn't increasing the number of playerKills when an enemy is killed. We need to adjust the spritesheet for the gunboat because it doesn't load properly. Also, other sprites loaded facing backwards... that should be an easy fix though. Despite all that, we have enemies and we can kill them! Progress!