Treasure Randomly Spawns
Every so often, treasure will blow in from where the wind is coming from. I also tried to make a function that prevents the player from spawning dead center when there is an island there, but it recurses infinitely because for some reason every rectangle's x and y values are 0 no matter where it is in the game.