Mostly implemented score
Score is almost completely working.  I also fixed a bug that maed not every enemy wrap around the world. Fruitlessly tried to get the player-not-spawning-in-the-middle code working, and also struggled (and failed) to get score working for treasure that is spawned by killed enemies. Added score indicator to the top of the screen. Fixed typo tha tkept giving a warning that tentacles.png couldn't load.