Added boarding pirate mechanic
It's the parrot sprite for now (my photoshop license expired or I'd have already made it), but if the player has a  pirate, it can be shot dead ahead for an instant kill. We may want to tweak its range and determine whether we want it to be heat-seeking or just shoot straight ahead. Still need to make it collect treasure, too.