Finished implementing score and treasure collection
I created a new class to allow easier global access of some of our variables- I think it may be sueful in solving the Kraken problem, too. As of now, dtopped treasure is properly collected and adds to the score, and using the boarding pirate attack adds to the players' score too. I als omade some minor changes to improve powerup spawning.