Gunboats sort of shoot
Only one of them does it at a time, so I think there's a problem with how I implemented EnemyWeapons (perhaps we need an array/group of enemies and an array/group of weapon arrays), and they only shoot north, but two enemy weapons fire. I also adjusted the balance for the player's firing range so it takes many more kills to get to max range. Added comments to storage.js