Implemented Basic version of MOAB
I forgot to add sails to the upwind sprites on the spritesheet, so those are on now. I also added the code for the MOAB boss, which works as anticipated. We need to figure out how to make more complicated firing. patterns and how to make individual weapons fire from different places on the sprite. Also reduced max number of islands to 6, which seems good for now.