Improved on MOAB
Imrpoved on MOAB firing patterns. Added a function to get ships to turn away from islands when they've hit one, but it noly works sometimes. Shrunk the size of the rocket image because it was way too big. For some reason, it was using the tiny original image for a long time instead of the medium one, so I thought I needed to make it bigger than I did...hopefully this time the size looks good