Potentially patched memory leak, small changes
Fixed a bug where the wind always came from the East on the first wave, added two functions to remove dead sprites (which is hopefully what was causing the memory leak), and commented out a ton of tests. Also made a change to the normal enemy AI to make them not run away as soon.