Fixed pirate facts text
I tried a bunch of stuff, but the easiest thing to do was just to put the second half of the facts in a second text file. The story about Jack Rackham stealing a Spanish ship while they blew up his old one had to go because it was way too long, but everything else was good. If any changes are made to either one, it's important to make sure that the lines match up properly.

Also added blood and waer splash when the Kraken gets hit, which I guess I had forgotten before.