Added actual song, started to add cannon sounds
The actual song plays now instead of the placeholder, and I added some code to try and figure out cannon firing sounds, but because of how fast all of the cannons fire it sounds like an A-10. The code that plays cannon sounds is commented out for now.
As per some of the tester feedback, I separated the pirate facts from the rest of the information on the game over screen.